finished override

+++ src/shapes/Rectangle.java
@@ -1,10 +1,10 @@
 package shapes;
 
 public class Rectangle {
-    protected int length;
-    protected int width;
+    protected double length;
+    protected double width;
 
-    public Rectangle(int length, int width){
+    public Rectangle(double length, double width){
         this.length = length;
         this.width = width;
     }
@@ -14,11 +14,11 @@ public class Rectangle {
     }
 
 
-    public int getArea(){
+    public double getArea(){
         return this.length * this.width;
     }
 
-    public int getPerimeter(){
+    public double getPerimeter(){
         return (this.width * 2) + (this.length * 2);
     }
